OM class

This is the format for all output for the OM functions. The output is a list with meta A list with elements that describe the model used.

version: Version of WRAP package used.

func: Name of function that produced the grid.

call: Exact call used to run model.

sim.seed: The random seed used for the simulation to replicate the results .Random.seed <- sim.seed before rerunning the OM function.

entries for all the arguments to the function (function dependent)

covariates: The covariates included in the model.

grid.dimensions: c(nrow, ncol, n cells).

grid.resolution: c(x, y)

grid.extent: c(xmin, xmax, ymin, ymax)

grid.crs: grid coordinate system

grid.unit Eg. "degree" Add units for clarity. NA if grid is simulated for example a 20x20 grid.

time: c(start date, end date). For example c(2001, 2101)

time.unit = "year" for yearly.

grid A data frame with a row for each grid square. The data frame has the following columns. Other columns can be added but at minimum the following are provided:

lat

lon

year

pres Presence/absence as 0/1

suitability

columns for the covariates (one column for each)

covariates: The covariates included in the model. The number of covariates will vary by model.

abundance