Simulated World basic

Function to simulate species distribution and abundance with a linearly increasing temperature. The argument temp_diff specifies the range of temperature for the beginning and ending years (year 1 and year 100).

SimulateWorld(
  temp_diff = c(1, 3, 5, 7),
  temp_spatial = c("simple", "matern"),
  PA_shape = c("logistic", "logistic_prev", "linear"),
  abund_enviro = c("lnorm_low", "lnorm_high", "poisson"),
  covariates = c("temp"),
  grid.size = 20,
  n.year = 100,
  start.year = 2000,
  response.curve = list(temp = c(fun = "dnorm", mean = 4, sd = 1))
)

Arguments

temp_diff
specifies min and max temps at year 1 and year 100 (e.g. temp_diff=c(1,3,5,7) means year 1 varies from 1-3C and year 100 from 5-7C)

temp_spatial
specifies whether we have "simple" linear temp distbn (SB) or added "matern" variation (EW)

PA_shape
specifies how enviro suitability determines species presence-absence. takes values of "logistic" (SB original), "logistic_prev" (JS, reduces knife-edge), "linear" (JS, reduces knife edge, encourages more absences, currently throws errors)

abund_enviro
specifies abundance if present, can be "lnorm_low" (SB original), "lnorm_high" (EW), or "poisson" (JS, increases abundance range)

covariates
Currently only "temp" is allowed in this function

grid.size
Default is a 20x20 grid

n.year
Number of years to simulate. Default is 100.

start.year
For showing results choose the start year.

response.curve
The response curve passed to `virtualspecies::formatFunctions`. All covariates in `covariates` argument must have a response curve specified.

Examples

# use defaults
data <- SimulateWorld()$grid
# plot time-series of total catch in observed years
plot(aggregate(abundance~year,data[data$year<=2010,], FUN="sum"),type="l",ylab="Abundance")
# plot time-series of total catch in forecast years
plot(aggregate(abundance~year,data[data$year>2010,], FUN="sum"),type="l", ylab="Abundance")