Plots COG from SDM

plots the center of gravity versus time

plot_cog(x, ...)

Arguments

x
An OM object output by one of the simulation functions.

...
SDMs to compare. Must be an SDM object output from one of the fitting functions (gam_sdm, brt_sdm, or mlp_sdm) or POM object output by predict.OM.

Details

The SDMs which will be compared are SDM objects and have the `start.forecast.year` as part of their meta data (in $meta). The COG will be plotted for the training data (to which the SDM was fit) and the test data (after `start.forecast.year`). If `start.forecast.year` is greater than the last year of data, then the forecast years are not plotted (since there is not data for those years).

Examples

sim <- SimulateWorld(start.year=2015, n.year=20)
mlp.fit <- mlp_sdm(sim, "temp")
plot_cog(sim, mlp.fit)