Predict for OM object

Does a hindcast and forecast with OM model. Optionally previously fitted SDMs can passed in to avoid having to refit SDMs. If abund_enviro is log-normal, then the prediction requires a presence model and an abundance model. If these are not passed in, they will be estimated. If abund_enviro is poisson, then only the abundance model is needed. Both a hindcast and a forecast will be returned. If model is "gam", then 95

# S3 method for OM
predict(
  object,
  model = c("gam", "brt", "mlp"),
  newdata = NULL,
  start.forecast.year = 2021,
  p.sdm = NULL,
  a.sdm = NULL,
  alpha = 0.6826895,
  silent = FALSE,
  ...
)

Arguments

object
OM object

model
Type of SDM. gam, brt or mlp

newdata
(optional) Optional new data frame over which to do the prediction.

start.forecast.year
The years equal or less will be used for fitting and the years greater than are the forecasted years.

p.sdm
(optional) Previously fitted presence model

a.sdm
(optional) Previously fitted abundance model

alpha
The CIs alpha. Default is 1 sigma (alpha = 68%)

silent
No output printed.

...
Any extra parameters for the sdm fitting functions, for example k for gam_sdm

Value

Invisibly returns a data frames with the prediction. If newdata is not passed in, then the hindcast and forecast prediction is returned. The "type" column has "hindcast" and "forecast" to indicate which is which. This data frame is the same shape as object$grid except the extra column "type". The first 6 rows of the prediction are printed unless silent=TRUE. If "newdata" is passed in then the prediction for the new data only is returned. "newdata" must be a data frame with the required covariates.