Plot yearly COG

Plots the center of gravity versus year for objects of class OM (returned by simulations functions), SDM (returned by fitting functions) and/or POM (returned by predict function).

plot_cog(x, ...)

Arguments

x
Object to plot. Can be a OM or POM object.

...
more objects to compare. Can be a SDM, POM or OM.

Details

If any of the objects to be compared are SDM objects, then x must be a OM object and predicts will be made for the grid points in the OM object.

If the objects to be compared are SDM objects and have the `start.forecast.year` as part of their meta data (in $meta), a reference line will be plotted showing the training versus test years. If `start.forecast.year` is greater than the last year of data, then no reference line will be plotted.

Examples

sim <- SimulateWorld(start.year=2015, n.year=20)
mlp.fit <- mlp_sdm(sim, "temp")
plot_cog(sim, mlp.fit)